(feat) Put table into separate component
